core/views.py
from django.shortcuts import render

from rest_framework.viewsets import ModelViewSet

from core.models import Categoria, Carro, Marca
from core.serializers import CategoriaSerializer, CarroSerializer, MarcaSerializer

class CategoriaViewSet(ModelViewSet):
    queryset = Categoria.objects.all()
    serializer_class = CategoriaSerializer


class CarroViewSet(ModelViewSet):
    queryset = Carro.objects.all()
    serializer_class = CarroSerializer


class MarcaViewSet(ModelViewSet):
    queryset = Marca.objects.all()
    serializer_class = MarcaSerializer
